implemented preloading and fixed jest exit issue
