+++ server/package.json
@@ -2,11 +2,11 @@
   "name": "asti-memories",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "server.js",
   "scripts": {
-    "nodemon": "nodemon --watch app.js",
+    "nodemon": "nodemon --watch server.js",
     "test": "NODE_ENV=test jest --detectOpenHandles",
-    "start": "node app.js"
+    "start": "node server.js"
   },
   "repository": {
     "type": "git",
@@ -35,7 +35,7 @@
     "express-jwt": "^8.4.1",
     "jest": "^29.7.0",
     "jsonwebtoken": "^9.0.0",
-    "nodemon": "^1.14.9",
+    "nodemon": "^3.0.2",
     "pg": "^8.10.0",
     "pg-hstore": "^2.3.4",
     "postgres": "^3.3.4",
