+++ server/server.js
@@ -2,7 +2,7 @@ const router = require("./router");
 const {sequelize} = require('./models');
 const env = process.env.NODE_ENV || "development"
 const port = env === "test" ? 8080 : 3000
-
+const app = require("./app")
 
 const connectToDB = async () => {
     console.log("connectToDB called")
@@ -15,19 +15,14 @@ const connectToDB = async () => {
     }
 }
 
-const startServer = (app) => {
-    (async () => {
-        await connectToDB();
-    })();
-    router(app);
-    app.listen(port, (err) => {
-        if(err) {
-            throw err;
-        }
-        console.log(`Listening on port ${port}`);
-    });
-}
-
-module.exports = {startServer};
+(async () => {
+    await connectToDB();
+})();
+app.listen(port, (err) => {
+    if(err) {
+        throw err;
+    }
+    console.log(`Listening on port ${port}`);
+});
